requirements.txt
Django==3.1.2
pytz==2018.5
djangorestframework~=3.12.1
django-filter~=2.4.0
pytest~=6.2.4